requirement.txt
pyzk
mysql-connector-python